Chnage datatype from int to long for identifier variable
SVN-Revision: 83

+++ WEB-INF/src/edu/wustl/catissuecore/domain/StorageType.java
@@ -22,7 +22,7 @@ public class StorageType implements java.io.Serializable
 	/**
      * System generated unique identifier.
      */
-	protected Integer systemIdentifier;
+	protected Long systemIdentifier;
 	
 	/**
      * Text name assigned to the container type
@@ -43,10 +43,10 @@ public class StorageType implements java.io.Serializable
      * Returns System generated unique identifier.
      * @return Integer System generated unique identifier.
      * @see #setSystemIdentifier(Integer)
-     * @hibernate.id name="systemIdentifier" column="IDENTIFIER" type="int" length="30"
+     * @hibernate.id name="systemIdentifier" column="IDENTIFIER" type="long" length="30"
      * unsaved-value="null" generator-class="native" 
      * */
-	public Integer getSystemIdentifier()
+	public Long getSystemIdentifier()
 	{
 		return systemIdentifier;
 	}
@@ -56,7 +56,7 @@ public class StorageType implements java.io.Serializable
      * @param systemIdentifier identifier of the Storagetype to be set.
      * @see #getSystemIdentifier()
      * */
-	public void setSystemIdentifier(Integer systemIdentifier)
+	public void setSystemIdentifier(Long systemIdentifier)
 	{
 		this.systemIdentifier = systemIdentifier;
 	}
